+++ psyneulink/core/components/functions/__init__.py
@@ -7,6 +7,7 @@ from . import selectionfunctions
 from . import integratorfunctions
 from . import distributionfunctions
 from . import objectivefunctions
+from . import optimizationfunctions
 
 from .function import *
 from .userdefinedfunction import *
@@ -17,6 +18,7 @@ from .selectionfunctions import *
 from .integratorfunctions import *
 from .distributionfunctions import *
 from .objectivefunctions import *
+from .optimizationfunctions import *
 
 __all__ = list(function.__all__)
 __all__.extend(userdefinedfunction.__all__)
@@ -27,3 +29,4 @@ __all__.extend(selectionfunctions.__all__)
 __all__.extend(integratorfunctions.__all__)
 __all__.extend(distributionfunctions.__all__)
 __all__.extend(objectivefunctions.__all__)
+__all__.extend(optimizationfunctions.__all__)
