+++ psyneulink/core/components/mechanisms/adaptive/control/optimizationcontrolmechanism.py
@@ -249,7 +249,9 @@ import itertools
 import numpy as np
 import typecheck as tc
 
-from psyneulink.core.components.functions.function import Function_Base, ModulationParam, OBJECTIVE_FUNCTION, SEARCH_FUNCTION, SEARCH_SPACE, SEARCH_TERMINATION_FUNCTION, _is_modulation_param, is_function_type
+from psyneulink.core.components.functions.function import Function_Base, ModulationParam, _is_modulation_param, is_function_type
+from psyneulink.core.components.functions.optimizationfunctions import OBJECTIVE_FUNCTION, SEARCH_FUNCTION, \
+    SEARCH_SPACE, SEARCH_TERMINATION_FUNCTION
 from psyneulink.core.components.mechanisms.adaptive.control.controlmechanism import ControlMechanism
 from psyneulink.core.components.mechanisms.processing.objectivemechanism import MONITORED_OUTPUT_STATES, ObjectiveMechanism
 from psyneulink.core.components.states.modulatorysignals.controlsignal import ControlSignal, ControlSignalCosts
